Fenrir/pyproject.toml
# This file defines the project metadata and dependencies for Fenrir.
# It is managed by Poetry. See https://python-poetry.org/ for documentation.

[tool.poetry]
name = "fenrir-scanner"
version = "2.0.0"
description = "A CLI-based multi-module security scanner for penetration testing."
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "fenrir"}]

# This section tells Poetry how to create a runnable script for your application.
[tool.poetry.scripts]
fenrir = "fenrir.cli:main"

[tool.poetry.dependencies]
# Pinned Python version to ensure compatibility with all dependencies (e.g., nvdlib).
python = ">=3.10, <3.13" 
python-nmap = "^0.7.1"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
scapy = "^2.5.0"
python-dotenv = "^1.0.1"
colorama = "^0.4.6"
Pillow = "^10.4.0"
httpx = "^0.27.0"
aiodns = "^3.2.0"
python-whois = "^0.8.0"
webtech = "^1.2.1"
asyncio-nmap = "^0.1.1"
nvdlib = "^1.1.1"
paramiko = "^3.4.0"
bleak = "^0.21.1"
paho-mqtt = "^2.1.0"
androguard = "^4.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
black = "^24.4.2"
ruff = "^0.4.2"

# --- Tool Configuration ---

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.venv
  | _build
  | build
  | dist
  | data
)/
'''

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "C90", "N"]
ignore = []

exclude = [
    ".eggs",
    ".git",
    ".venv",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "venv",
    "data",
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
#
